vue3: 完成菜单管理改造

+++ wk-mini/src/main/java/com/budwk/app/sys/controllers/sys/SysMenuController.java
@@ -6,6 +6,7 @@ import com.budwk.app.sys.services.SysAppService;
 import com.budwk.app.sys.services.SysMenuService;
 import com.budwk.app.sys.services.SysRoleService;
 import com.budwk.app.sys.services.SysUserService;
+import com.budwk.starter.common.constant.GlobalConstant;
 import com.budwk.starter.common.openapi.annotation.*;
 import com.budwk.starter.common.openapi.enums.ParamIn;
 import com.budwk.starter.common.page.Pagination;
@@ -73,9 +74,9 @@ public class SysMenuController {
             implementation = Pagination.class
     )
     @SaCheckPermission("sys.manage.menu")
-    public Result<?> list(@Param("name") String name,@Param("appId") String appId,@Param("href") String href) {
+    public Result<?> list(@Param("name") String name, @Param("appId") String appId, @Param("href") String href) {
         Cnd cnd = Cnd.NEW();
-        cnd.and("appId","=",appId);
+        cnd.and("appId", "=", appId);
         if (Strings.isNotBlank(name)) {
             cnd.and("name", "like", "%" + name + "%");
         }
@@ -253,9 +254,9 @@ public class SysMenuController {
             return Result.error(ResultCode.NULL_DATA_ERROR);
         }
         req.setAttribute("_slog_msg", menu.getName());
-        // if (Strings.sNull(menu.getPath()).startsWith("0001")) {
-        //     return Result.error("系统菜单禁止操作");
-        // }
+        if (!GlobalConstant.DEFAULT_SYSADMIN_LOGINNAME.equals(SecurityUtil.getUserLoginname()) && Strings.sNull(menu.getPath()).startsWith("0001")) {
+            return Result.error("系统菜单禁止操作");
+        }
         sysMenuService.deleteAndChild(menu);
         return Result.success();
     }
